desktop/backend/python/requirements.txt
pyautogui==0.9.53
psutil==7.2.0
Pillow==12.0.0
mss==10.1.0
pynput==1.8.1